requirements.txt
click==8.1.7
pytorch-lightning==2.4.0
hydra-core==1.3.2
clearml==1.16.4
torchmetrics==1.4.1
timm==0.9.7
pandas==2.2.2
scikit-learn==1.5.1
matplotlib==3.9.2
seaborn==0.13.2
